chore(posts): move PostsEntity to entity/ and PostsRepository to repository/
